# voya-agent-openapi.yaml
openapi: 3.0.0
info:
  title: Voya Agent API
  version: 1.0.0
  description: >
    Complete API specification for **Voya Agent**, an AI-powered travel assistant
    that helps users plan trips, find tours, and manage conversations with a chat interface.
servers:
  - url: https://voya-agent-j9zu.onrender.com
paths:
...
  /api/conversations/{conversation_id}/:
    get:
      summary: Get Conversation Details
      description: Get all messages and metadata for a specific conversation.
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema: { type: integer }
        - name: session_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Conversation details.
...